+++ src/main/java/ippo/assignment2/App.java
@@ -1,24 +1,28 @@
 package ippo.assignment2;
 
-import com.sun.prism.ResourceFactoryListener;
 import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 
-import java.io.File;
 import java.io.IOException;
-import java.util.ResourceBundle;
+import java.io.InputStream;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 import ippo.assignment2.controllers.MainController;
+import ippo.assignment2.utils.Properties;
 
 /**
  * @since 0.1.2
  */
 public class App extends javafx.application.Application {
+
+    private final String  mainControllerViewerPath =  "/fxml/MainControllerViewer.fxml";
+    private Properties properties = new Properties();
+
+    public App() throws IOException {
+    }
 
     /**
      *
@@ -38,16 +42,11 @@ public class App extends javafx.application.Application {
 
         try {
 
-            String viewerFxml = "/fxml/MainControllerViewer.fxml";
-
-
-            FXMLLoader fxmlLoader = new FXMLLoader(getClass().getResource(viewerFxml));
-            AnchorPane page = (AnchorPane) fxmlLoader.load();
-
+            FXMLLoader loader = new FXMLLoader(getClass().getResource(this.mainControllerViewerPath));
+            AnchorPane page =  (AnchorPane) loader.load(getClass().getResource(mainControllerViewerPath));
             Scene scene = new Scene(page);
             stage.setScene(scene);
-            MainController mainController = fxmlLoader.getController();
-
+            MainController mainController = loader.getController();
             stage.show();
 
         } catch (IOException ex) {