+++ src/main/java/ippo/assignment2/singletons/PlayerSingleton.java
@@ -1,5 +1,0 @@
-package ippo.assignment2.singletons;
-
-public class PlayerSingleton {
-
-}